fix(plutus): consolidate inventory pages into tabs
